grpc_test_final/other_en/requirements.txt
grpcio==1.48.2
grpcio-tools==1.48.2
protobuf==3.20.3
etcd3
requests